added sign up + fixed login + little redesign

+++ backend/controllers/notesController.js
@@ -1,15 +1,16 @@
 const mongoose = require("mongoose");
 const Note = require("../models/note");
+const User = require("../models/user");
 
 const getAllNotes = async (req, res) => {
   try {
-    const notes = await Note.find({}).populate("user", {
+    const notes = await Note.find({ user: req.user }).populate("user", {
       username: 1,
       name: 1,
     });
     return res.status(200).json(notes);
   } catch (error) {
-    return res.status(500).json({ error: "Failed to fetch all notes" });
+    return res.status(500).json({ error: "Failed to fetch notes" });
   }
 };
 
@@ -26,6 +27,10 @@ const getNoteById = async (req, res) => {
       return res.status(404).json({ error: "Note not found" });
     }
 
+    if (note.user.toString() !== req.user) {
+      return res.status(403).json({ error: "Access denied" });
+    }
+
     return res.status(200).json(note);
   } catch (error) {
     return res.status(500).json({ error: "Failed to fetch note" });
@@ -40,8 +45,19 @@ const createNewNote = async (req, res) => {
   }
 
   try {
-    const newNote = new Note({ title, content });
+    const user = await User.findById(req.user);
+
+    const newNote = new Note({
+      title,
+      content,
+      user: user._id,
+    });
+
     const savedNote = await newNote.save();
+
+    user.notes = user.notes.concat(savedNote._id);
+    await user.save();
+
     return res.status(201).json(savedNote);
   } catch (error) {
     return res.status(500).json({ error: "Failed to create new note" });
@@ -56,15 +72,19 @@ const updateNote = async (req, res) => {
   }
 
   try {
-    const updatedNote = await Note.findByIdAndUpdate(id, req.body, {
-      new: true,
-      runValidators: true,
-      context: "query",
-    });
-
-    if (!updatedNote) {
+    const note = await Note.findById(id);
+    if (!note) {
       return res.status(404).json({ error: "Note not found" });
     }
+
+    if (note.user.toString() !== req.user) {
+      return res.status(403).json({ error: "Access denied" });
+    }
+
+    note.title = req.body.title || note.title;
+    note.content = req.body.content || note.content;
+
+    const updatedNote = await note.save();
 
     return res.status(200).json(updatedNote);
   } catch (error) {
@@ -80,10 +100,16 @@ const deleteNote = async (req, res) => {
   }
 
   try {
-    const deletedNote = await Note.findByIdAndDelete(id);
-    if (!deletedNote) {
+    const note = await Note.findById(id);
+    if (!note) {
       return res.status(404).json({ error: "Note not found" });
     }
+
+    if (note.user.toString() !== req.user) {
+      return res.status(403).json({ error: "Access denied" });
+    }
+
+    await Note.findByIdAndDelete(id);
 
     return res.status(204).end();
   } catch (error) {
